[WIP] Account linking (#18)
* Work on account linking.

* Account linking now works
Also updated jda dependency

* Bit of cleanup/improvements
Renamed AbstractConfig to BaseConfig - was using abstract methods in early tests, not anymore.

Closes #3

+++ src/main/java/io/github/nucleuspowered/phonon/discord/command/CommandListener.java
@@ -7,6 +7,7 @@ import io.github.nucleuspowered.phonon.discord.DiscordCommandSource;
 import io.github.nucleuspowered.phonon.modules.core.CoreModule;
 import io.github.nucleuspowered.phonon.modules.core.config.CoreConfig;
 import io.github.nucleuspowered.phonon.modules.core.config.CoreConfigAdapter;
+import net.dv8tion.jda.core.entities.PrivateChannel;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.core.hooks.ListenerAdapter;
 import org.spongepowered.api.command.args.ArgumentParseException;
@@ -54,9 +55,8 @@ public class CommandListener extends ListenerAdapter {
     public void onMessageReceived(MessageReceivedEvent event) {
         CoreConfig config = this.phononPlugin.getConfigAdapter(CoreModule.ID, CoreConfigAdapter.class).get().getNodeOrDefault();
 
-        if (event.getMessage().getRawContent().startsWith(config.getPrefix()) && !event.getMessage().getAuthor().isBot() &&
-                event.getTextChannel() != null) {
-            DiscordCommandSource source = new DiscordCommandSource(event.getAuthor(), event.getTextChannel());
+        if (event.getMessage().getRawContent().startsWith(config.getPrefix()) && !event.getMessage().getAuthor().isBot()) {
+            DiscordCommandSource source = new DiscordCommandSource(event.getAuthor(), event.getChannel());
 
             int end = event.getMessage().getRawContent().indexOf(" ");
             String commandName;
@@ -91,13 +91,16 @@ public class CommandListener extends ListenerAdapter {
                     }
 
                     if (args == null) {
-                        event.getChannel().sendMessage(message);
+                        event.getChannel().sendMessage(message).queue();
                     } else {
                         event.getChannel().sendMessage(message + System.lineSeparator() +
                                 config.getPrefix() + commandName + args.getUsage(source).toPlain()).queue();
                     }
                 }
-                event.getMessage().delete().queue();
+
+                if (!(event.getChannel() instanceof PrivateChannel)) {
+                    event.getMessage().delete().queue();
+                }
             }
         }
     }
